+++ src/XmlTools/XmlParser.java
@@ -1,4 +1,4 @@
-package controllers;
+package XmlTools;
 
 import java.io.File;
 import java.io.IOException;
@@ -8,11 +8,7 @@ import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Source;
-import javax.xml.transform.stream.StreamSource;
-import javax.xml.validation.Schema;
 import javax.xml.validation.SchemaFactory;
-import javax.xml.validation.Validator;
 import org.w3c.dom.Document;
 import org.xml.sax.SAXException;
 
@@ -41,8 +37,8 @@ public class XmlParser{
         return parser.parse(filePath);
     }
 
-    public  ValidatorSchema getVAlidatorSchema(File xsdFile) throws SAXException{
-        return ValidatorSchema.getValidator(sf.newSchema(xsdFile));
+    public  ValidatorXMLSchema getVAlidatorSchema(File xsdFile) throws SAXException{
+        return ValidatorXMLSchema.getValidator(sf.newSchema(xsdFile));
     }
 
 }
